requirements.txt
Django==5.0.3
djangorestframework==3.15.0

